firebase/functions/requirements.txt
# Firebase Cloud Functions dependencies
firebase-functions==0.4.0
firebase-admin==6.4.0

# Web scraping
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.1.0

# AI and vector DB
openai==1.12.0
pinecone-client==3.1.0

# Utilities
python-dotenv==1.0.1
